refactor: qabot 리팩토링
